Consensus: MOVEONLY: Move CValidationState from main consensus/validation
