+++ src/miner.cpp
@@ -10,6 +10,7 @@
 
 #include "amount.h"
 #include "consensus/consensus.h"
+#include "consensus/validation.h"
 #include "hash.h"
 #include "main.h"
 #include "masternode-sync.h"
