+++ src/rpc/blockchain.cpp
@@ -7,7 +7,9 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "checkpoints.h"
+#include "consensus/validation.h"
 #include "main.h"
+#include "primitives/transaction.h"
 #include "rpc/server.h"
 #include "sync.h"
 #include "util.h"
