+++ src/test/miner_tests.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "init.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "miner.h"
 #include "pubkey.h"
