+++ src/test/transaction_tests.cpp
@@ -6,12 +6,13 @@
 #include "data/tx_valid.json.h"
 
 #include "clientversion.h"
+#include "consensus/validation.h"
+#include "core_io.h"
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
 #include "script/script.h"
 #include "script/script_error.h"
-#include "core_io.h"
 
 #include <map>
 #include <string>
