+++ src/txmempool.cpp
@@ -9,6 +9,7 @@
 
 #include "clientversion.h"
 #include "consensus/consensus.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "streams.h"
 #include "util.h"
