+++ src/wallet.cpp
@@ -13,6 +13,7 @@
 #include "checkpoints.h"
 #include "coincontrol.h"
 #include "consensus/consensus.h"
+#include "consensus/validation.h"
 #include "kernel.h"
 #include "masternode-budget.h"
 #include "net.h"
