+++ src/walletdb.cpp
@@ -9,6 +9,7 @@
 #include "walletdb.h"
 
 #include "base58.h"
+#include "consensus/validation.h"
 #include "protocol.h"
 #include "serialize.h"
 #include "sync.h"
